all: move event package from p2p to common
Merges the go-ethereum event package (Feed, FeedOf, Subscription,
SubscriptionScope) into the existing common/event (Observers,
Notifier). No name collisions, no behavior change — files moved
verbatim, import paths rewritten.
