+++ cl/beacon/beaconevents/state_feed.go
@@ -1,7 +1,7 @@
 package beaconevents
 
 import (
-	ethevent "github.com/erigontech/erigon/p2p/event"
+	ethevent "github.com/erigontech/erigon/common/event"
 )
 
 type stateFeed struct {
