+++ execution/abi/bind/template.go
@@ -101,7 +101,7 @@ import (
 	"github.com/erigontech/erigon/execution/abi/bind"
 	"github.com/erigontech/erigon/common"
 	"github.com/erigontech/erigon/execution/types"
-	"github.com/erigontech/erigon/p2p/event"
+	"github.com/erigontech/erigon/common/event"
 )
 
 // Reference imports to suppress errors if they are not otherwise used.
